simple optimisations
